requirements.txt
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
chapa==0.1.2
crispy-bootstrap4==2024.10
Django==4.2.16
django-apscheduler==0.7.0
django-bootstrap-v5==1.0.11
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filter==24.3
django-formtools==2.5.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
environ==1.0
exceptiongroup==1.2.2
fontawesomefree==6.6.0
GDAL @ file:///croot/gdal-split_1734448174900/work/build/swig/python
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1651020388975/work
phonenumbers==9.0.3
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
